refactor: antispam escalation logging funcs
Remove `LoggingModule.logMemberSpamming` and
`LoggingModule.logMemberTimeout`. Move functionality to Antispam in
Automod module.
